backend/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
SQLAlchemy==2.0.42
sqlalchemy-cockroachdb==2.0.3
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
